requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
google-generativeai==0.3.2
python-multipart==0.0.6
sentence-transformers==2.2.2
